Remove `HeadlessCallbackHandler` by defaulting callback functions
